refactor: move out generic event interfaces into separate type file
