--- docs/code/enums/types_async_event_emitter.EventType.md
@@ -1,32 +1,0 @@
-[@algorandfoundation/algokit-utils](../README.md) / [types/async-event-emitter](../modules/types_async_event_emitter.md) / EventType
-
-# Enumeration: EventType
-
-[types/async-event-emitter](../modules/types_async_event_emitter.md).EventType
-
-## Table of contents
-
-### Enumeration Members
-
-- [AppCompiled](types_async_event_emitter.EventType.md#appcompiled)
-- [TxnGroupSimulated](types_async_event_emitter.EventType.md#txngroupsimulated)
-
-## Enumeration Members
-
-### AppCompiled
-
-• **AppCompiled** = ``"AppCompiled"``
-
-#### Defined in
-
-[src/types/async-event-emitter.ts:5](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/async-event-emitter.ts#L5)
-
-___
-
-### TxnGroupSimulated
-
-• **TxnGroupSimulated** = ``"TxnGroupSimulated"``
-
-#### Defined in
-
-[src/types/async-event-emitter.ts:4](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/async-event-emitter.ts#L4)
