+++ docs/code/enums/types_lifecycle_events.EventType.md
@@ -1,0 +1,32 @@
+[@algorandfoundation/algokit-utils](../README.md) / [types/lifecycle-events](../modules/types_lifecycle_events.md) / EventType
+
+# Enumeration: EventType
+
+[types/lifecycle-events](../modules/types_lifecycle_events.md).EventType
+
+## Table of contents
+
+### Enumeration Members
+
+- [AppCompiled](types_lifecycle_events.EventType.md#appcompiled)
+- [TxnGroupSimulated](types_lifecycle_events.EventType.md#txngroupsimulated)
+
+## Enumeration Members
+
+### AppCompiled
+
+• **AppCompiled** = ``"AppCompiled"``
+
+#### Defined in
+
+[src/types/lifecycle-events.ts:5](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/lifecycle-events.ts#L5)
+
+___
+
+### TxnGroupSimulated
+
+• **TxnGroupSimulated** = ``"TxnGroupSimulated"``
+
+#### Defined in
+
+[src/types/lifecycle-events.ts:4](https://github.com/algorandfoundation/algokit-utils-ts/blob/main/src/types/lifecycle-events.ts#L4)
