+++ src/index.ts
@@ -17,5 +17,5 @@ export * from './debugging'
 export * as indexer from './indexer-lookup'
 export * from './transaction'
 export { AlgorandClient } from './types/algorand-client'
-export * from './types/async-event-emitter'
 export * from './types/debugging'
+export * from './types/lifecycle-events'
