+++ src/types/composer.ts
@@ -5,8 +5,8 @@ import { TransactionSignerAccount } from './account'
 import { AlgoAmount } from './amount'
 import { APP_PAGE_MAX_SIZE } from './app'
 import { AppManager, BoxIdentifier, BoxReference } from './app-manager'
-import { EventType } from './async-event-emitter'
 import { Expand } from './expand'
+import { EventType } from './lifecycle-events'
 import { genesisIdIsLocalNet } from './network-client'
 import { Arc2TransactionNote, SendAtomicTransactionComposerResults, SendParams } from './transaction'
 import AtomicTransactionComposer = algosdk.AtomicTransactionComposer
